requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pyyaml>=6.0
tqdm>=4.65.0

# LLM APIs
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
google-genai
together>=1.0.0

# Optional (uncomment if needed)
# transformers
# huggingface-hub
# sentence_transformers
# accelerate
# deepspeed
# wandb
